finish styling, add custom error message
There was too much padding which caused the page to be too wide.

Secondly, the image did not cover the page which caused a scroll bar
to appear vertically. The image would shrink but now it's retaining
the height of the page when the width of the page changes.

input:focus border was changed but doesn't apply to the button. Added
another custom error for the phone number using the regex pattern in js.

Lastly, the odin logo, credit, and semi-transparent background was in an
absolute fixed position. Now, its position is relative to the background
image.